feat(examples): basic multi-arch client program setup
Moves the `example` client programs into the workspace by adding
conditional `main` and `panic_handler` functions, depending on the
target environment.

As a follow-up, we should add a nice proc macro that hides this
functionality so that client programs can easily be linted and checked
on the native development platform.