chore: remove unused BlobProvider trait implementation on OnlineBlobProvider
Not sure why this was here, as its a private function that wasnt used anywhere in the crate.

Also confused me for a while in the beginning, as I kept wondering why the BlobProvider trait (which is meant to be implemented by ClientProvider) was being implemented by a host struct (OnlineBlobProvider). In my mind client and host traits should be separate afaiu.